[Project] Rename folders to lower_snake_case to follow conventions
