+++ MediaElch.pro
@@ -261,13 +261,13 @@ SOURCES += src/main.cpp \
     src/scrapers/music/TvTunes.cpp \
     src/scrapers/music/UniversalMusicScraper.cpp \
     src/scrapers/tv_show/TheTvDb.cpp \
-    src/scrapers/tv_show/TheTvDb/ApiRequest.cpp \
-    src/scrapers/tv_show/TheTvDb/Cache.cpp \
-    src/scrapers/tv_show/TheTvDb/EpisodeLoader.cpp \
-    src/scrapers/tv_show/TheTvDb/EpisodeParser.cpp \
-    src/scrapers/tv_show/TheTvDb/Search.cpp \
-    src/scrapers/tv_show/TheTvDb/ShowLoader.cpp \
-    src/scrapers/tv_show/TheTvDb/ShowParser.cpp \
+    src/scrapers/tv_show/thetvdb/ApiRequest.cpp \
+    src/scrapers/tv_show/thetvdb/Cache.cpp \
+    src/scrapers/tv_show/thetvdb/EpisodeLoader.cpp \
+    src/scrapers/tv_show/thetvdb/EpisodeParser.cpp \
+    src/scrapers/tv_show/thetvdb/Search.cpp \
+    src/scrapers/tv_show/thetvdb/ShowLoader.cpp \
+    src/scrapers/tv_show/thetvdb/ShowParser.cpp \
     src/ui/movie_sets/MovieListDialog.cpp \
     src/ui/movie_sets/SetsWidget.cpp \
     src/settings/AdvancedSettings.cpp \
@@ -518,13 +518,13 @@ HEADERS  += Version.h \
     src/scrapers/music/TvTunes.h \
     src/scrapers/music/UniversalMusicScraper.h \
     src/scrapers/tv_show/TheTvDb.h \
-    src/scrapers/tv_show/TheTvDb/ApiRequest.h \
-    src/scrapers/tv_show/TheTvDb/Cache.h \
-    src/scrapers/tv_show/TheTvDb/EpisodeLoader.h \
-    src/scrapers/tv_show/TheTvDb/EpisodeParser.h \
-    src/scrapers/tv_show/TheTvDb/Search.h \
-    src/scrapers/tv_show/TheTvDb/ShowLoader.h \
-    src/scrapers/tv_show/TheTvDb/ShowParser.h \
+    src/scrapers/tv_show/thetvdb/ApiRequest.h \
+    src/scrapers/tv_show/thetvdb/Cache.h \
+    src/scrapers/tv_show/thetvdb/EpisodeLoader.h \
+    src/scrapers/tv_show/thetvdb/EpisodeParser.h \
+    src/scrapers/tv_show/thetvdb/Search.h \
+    src/scrapers/tv_show/thetvdb/ShowLoader.h \
+    src/scrapers/tv_show/thetvdb/ShowParser.h \
     src/ui/movie_sets/MovieListDialog.h \
     src/ui/movie_sets/SetsWidget.h \
     src/settings/AdvancedSettings.h \
