+++ src/scrapers/tv_show/TheTvDb.cpp
@@ -8,10 +8,10 @@
 #include "media_centers/KodiXml.h"
 #include "movies/Movie.h"
 #include "scrapers/movie/IMDB.h"
-#include "scrapers/tv_show/TheTvDb/Cache.h"
-#include "scrapers/tv_show/TheTvDb/EpisodeLoader.h"
-#include "scrapers/tv_show/TheTvDb/Search.h"
-#include "scrapers/tv_show/TheTvDb/ShowLoader.h"
+#include "scrapers/tv_show/thetvdb/Cache.h"
+#include "scrapers/tv_show/thetvdb/EpisodeLoader.h"
+#include "scrapers/tv_show/thetvdb/Search.h"
+#include "scrapers/tv_show/thetvdb/ShowLoader.h"
 #include "settings/Settings.h"
 #include "ui/main/MainWindow.h"
