+++ src/scrapers/tv_show/thetvdb/EpisodeParser.cpp
@@ -1,7 +1,7 @@
 #include "EpisodeParser.h"
 
 #include "globals/Helper.h"
-#include "scrapers/tv_show/TheTvDb/ApiRequest.h"
+#include "scrapers/tv_show/thetvdb/ApiRequest.h"
 #include "settings/Settings.h"
 #include "tv_shows/TvShowEpisode.h"
 #include <QJsonArray>
