+++ src/scrapers/tv_show/thetvdb/Search.h
@@ -2,7 +2,7 @@
 
 #include "globals/Globals.h"
 #include "globals/ScraperResult.h"
-#include "scrapers/tv_show/TheTvDb/ApiRequest.h"
+#include "scrapers/tv_show/thetvdb/ApiRequest.h"
 
 #include <QJsonObject>
 #include <QNetworkAccessManager>
