+++ test/scrapers/thetvdb/testShowLoader.cpp
@@ -2,8 +2,8 @@
 
 #include "test/mocks/settings/MockScraperSettings.h"
 
-#include "scrapers/tv_show/TheTvDb/EpisodeLoader.h"
-#include "scrapers/tv_show/TheTvDb/ShowLoader.h"
+#include "scrapers/tv_show/thetvdb/EpisodeLoader.h"
+#include "scrapers/tv_show/thetvdb/ShowLoader.h"
 
 #include "settings/Settings.h"
